asMatrixCols {BBmisc}	R Documentation

Extracts a named element from a list of lists.

Description

Extracts a named element from a list of lists.

Usage

asMatrixCols(xs, row.names, col.names)

asMatrixRows(xs, row.names, col.names)

Arguments

xs
[list]
A list of vectors of the same length.

row.names
[character | integer | NULL]
Row names of result. Default is to take the names of the elements of xs.

col.names
[character | integer | NULL]
Column names of result. Default is to take the names of the elements of xs.

Value

[matrix].

[Package BBmisc version 1.6 Index]