cFactor {BBmisc}	R Documentation

Combine multiple factors and return a factor.

Description

Note that function does not inherit from c to not change R semantics behind your back when this package is loaded.

Usage

cFactor(...)

Arguments

...
[factor]
The factors.

Value

[factor].

Examples

f1 <- factor(c("a", "b"))
f2 <- factor(c("b", "c"))
print(c(f1, f2))

## [1] 1 2 1 2

print(cFactor(f1, f2))

## [1] a b b c
## Levels: a b c

[Package BBmisc version 1.6 Index]