chunk {BBmisc}	R Documentation

Chunk elements of vectors into blocks of nearly equal size.

Description

In case of shuffling and vectors that cannot be chunked evenly, it is chosen randomly which levels / chunks will receive 1 element less. If you do not shuffle, always the last chunks will receive 1 element less.

Usage

chunk(x, chunk.size, n.chunks, shuffle = FALSE)

Arguments

x
[vector | list]
Vector or list to chunk.

chunk.size
[integer(1)]
Requested number of elements in each chunk. Cannot be used in combination with n.chunks. If x cannot be evenly chunked, some chunks will have 1 element less.

n.chunks
[integer(1)]
Requested number of chunks. Can not be used in combination with chunks.size

shuffle
[logical(1)]
Shuffle x? Default is FALSE.

Value

[unnamed list] of chunks.

Examples

xs <- 1:10
chunk(xs, chunk.size=3)

## [[1]]
## [1] 1 2 3
## 
## [[2]]
## [1] 4 5 6
## 
## [[3]]
## [1] 7 8
## 
## [[4]]
## [1]  9 10

chunk(xs, n.chunks=2)

## [[1]]
## [1] 1 2 3 4 5
## 
## [[2]]
## [1]  6  7  8  9 10

chunk(xs, n.chunks=2, shuffle=TRUE)

## [[1]]
## [1]  1  2  4  8 10
## 
## [[2]]
## [1] 3 5 6 7 9

[Package BBmisc version 1.6 Index]