clearFunctionEnv {BBmisc}	R Documentation

Clear the environment of a function

Description

Clear the environment of a function

Usage

  clearFunctionEnv(f, inplace = FALSE)

Arguments

f
[function]
Function whose environment to clear.

inplace
[logical(1)]
If TRUE, remove all object from the function's environment. Assign a new one and let the garbage collector take care otherwise. If this is environment is the global environment, a new one will be assigned either way. Default is FALSE.

Value

Function with empty environment, parent is .GlobalEnv.

Examples

get_f = function() { x = 1; function() x }
f = get_f()
print(ls(environment(f)))

## [1] "x"

f = clearFunctionEnv(f)
print(ls(environment(f)))

## character(0)

[Package BBmisc version 1.5 Index]