convertIntegers {BBmisc}	R Documentation

Conversion for integer vector.

Description

Convert numeric vector to integer vector if the numeric vector fully represents an integer vector, e.g. c(1, 5) to c(1L, 5L). Otherwise the argument is returned unchanged.

Usage

convertIntegers(x)

Arguments

x
[any]
Argument.

Value

Either an integer vector if conversion was done or x unchanged.

Examples

str(convertIntegers(1.0))

##  int 1

str(convertIntegers(1.3))

##  num 1.3

str(convertIntegers(c(1.0, 2.0)))

##  int [1:2] 1 2

str(convertIntegers("foo"))

##  chr "foo"

[Package BBmisc version 1.6 Index]