convertListOfRowsToDataFrame {BBmisc}	R Documentation

Convert a list of row-vectors of equal structure to a data.frame.

Description

Convert a list of row-vectors of equal structure to a data.frame.

Usage

convertListOfRowsToDataFrame(rows,
  strings.as.factors = default.stringsAsFactors(), row.names, col.names)

Arguments

rows
[list]
List of rows. Each row is a list or vector of the same structure. That means all rows must have the same length and all corresponding elements must have the same class.

strings.as.factors
[logical(1)]
Convert character columns to factors? Default is default.stringsAsFactors().

row.names
[character(nrow) | NULL]
Row names for result. NULL implies no row names are set. By default the names of the list rows are taken.

col.names
[character(ncol)]
Column names. Default the are the names of the elements of the first row, or, if this row has no element names, “V1”, “V2”, and so on, is used.

Value

[data.frame].

Examples

convertListOfRowsToDataFrame(list(list(x=1, y="a"), list(x=2, y="b")))

##   x y
## 1 1 a
## 2 2 b

[Package BBmisc version 1.6 Index]