convertToShortString {BBmisc}	R Documentation

Converts any R object to a descriptive string so it can be used in messages.

Description

Atomics: If of length 0 or 1, they are basically printed as they are. Numerics are formated with num.format. If of length greater than 1, they are collpsed witd “,” and clipped. so they do not become excessively long.

Usage

convertToShortString(x, num.format = "%.4g", clip.len = 15L)

Arguments

x
[any]
The object.

num.format
[character(1)]
Used to format numerical scalars via sprintf. Default is “%.4g”.

clip.len
[integer(1)]
Used clip atomic vectors via clipString. Default is 15.

Details

All others: Currently, only their class is simply printed like “<data.frame>”.

Lists: The mechanism above is applied (non-recursively) to their elements. The result looks like this: “a=1, <unamed>=2, b=<data.frame>, c=<list>”.

Value

[character(1)].

Examples

convertToShortString(list(a=1, b=NULL, "foo", c=1:10))

## [1] "a=1, b=NULL, <unnamed>=foo, c=1,2,3,4,5,6,..."

[Package BBmisc version 1.6 Index]