do.call2 {BBmisc}	R Documentation

Execute a function call similar to do.call.

Description

This function is supposed to be a replacement for do.call in situations where you need to pass big R objects. Unlike do.call, this function allows to pass objects via ... to avoid a copy.

Usage

do.call2(fun, ..., .args = list())

Arguments

fun
[character(1)]
Name of the function to call.

...
[any]
Arguments to fun. Best practice is to specify them in a key=value syntax.

.args
[list]
Arguments to fun as a (named) list. Will be passed after arguments in .... Default is list().

Value

Return value of fun.

Examples

## Not run: 
##D   library(microbenchmark)
##D   x = 1:1e7
##D   microbenchmark(do.call(head, list(x, n=1)), do.call2("head", x, n=1))
## End(Not run)

[Package BBmisc version 1.6 Index]