dropNamed {BBmisc}	R Documentation

Drop named elements of an object.

Description

Drop named elements of an object.

Usage

dropNamed(x, drop)

Arguments

x
[any]
Object to drop named elements from. For a matrix or a data frames this function drops named columns via the second argument of the binary index operator [,]. Otherwise, the unary index operator [] is used for dropping.

drop
[character]
Names of elements to drop.

Value

Subset of object of same type as x. The object is not simplified, i.e, no dimensions are dropped as [,,drop = FALSE] is used.

[Package BBmisc version 1.6 Index]