getMaxIndexOfRows {BBmisc}	R Documentation

Find row- or columnwise the index of the maximal / minimal element in a matrix.

Description

getMaxIndexOfRows returns the index of the maximal element of each row. getMinIndexOfRows returns the index of the minimal element of each row. getMaxIndexOfCols returns the index of the maximal element of each col. getMinIndexOfCols returns the index of the minimal element of each col. If a corresponding vector (row or col) is empty, possibly after NA removal, -1 is returned as index.

Usage

getMaxIndexOfRows(x, ties.method = "random", na.rm = FALSE)

getMinIndexOfRows(x, ties.method = "random", na.rm = FALSE)

getMaxIndexOfCols(x, ties.method = "random", na.rm = FALSE)

getMinIndexOfCols(x, ties.method = "random", na.rm = FALSE)

Arguments

x
[matrix(n,m)]
Numerical input matrix.

ties.method
[character(1)]
How should ties be handled? Possible are: “random”, “first”, “last”. Default is “random”.

na.rm
[logical(1)]
If FALSE, NA is returned if an NA is encountered in x. If TRUE, NAs are disregarded. Default is FALSE

Value

[integer(n)].

Examples

x <- matrix(runif(5 * 3), ncol=3)
print(x)

##        [,1]    [,2]   [,3]
## [1,] 0.6061 0.30158 0.6895
## [2,] 0.9511 0.14420 0.3536
## [3,] 0.6121 0.67093 0.8442
## [4,] 0.2112 0.32374 0.8455
## [5,] 0.8051 0.01175 0.1470

print(getMaxIndexOfRows(x))

## [1] 3 1 3 3 1

print(getMinIndexOfRows(x))

## [1] 2 2 1 1 2

[Package BBmisc version 1.6 Index]