isSubset {BBmisc}	R Documentation

Check subset relation on two vectors.

Description

Check subset relation on two vectors.

Usage

isSubset(x, y, strict = FALSE)

Arguments

x
[vector]
Source vector.

y
[vector]
Vector of the same mode as x.

strict
[logical(1)]
Checks for strict/proper subset relation.

Value

[logical(1)] TRUE if each element of x is also contained in y, i. e., if x is a subset of y and FALSE otherwise.

[Package BBmisc version 1.6 Index]