isSuperset {BBmisc}	R Documentation

Check superset relation on two vectors.

Description

Check superset relation on two vectors.

Usage

isSuperset(x, y, strict = FALSE)

Arguments

x
[vector]
Source vector.

y
[vector]
Vector of the same mode as x.

strict
[logical(1)]
Checks for strict/proper superset relation.

Value

[logical(1)] TRUE if each element of y is also contained in x, i. e., if y is a subset of x and FALSE otherwise.

[Package BBmisc version 1.6 Index]