isValidName {BBmisc}	R Documentation

Can some strings be used for column or list element names without problems?

Description

Can some strings be used for column or list element names without problems?

Usage

isValidName(x, unique = TRUE)

Arguments

x
[character]
Character vector to check.

unique
[logical(1)]
Should the names be unique? Default is TRUE.

Value

[logical]. One Boolean entry for each string in x. If the entries are not unique and unique is enabled, the first duplicate will be FALSE.

[Package BBmisc version 1.6 Index]