lib {BBmisc}	R Documentation

A wrapper for library.

Description

Tries to load packages. If the packages are not found, they will be installed from the default repository. This function is intended for use in interactive sessions and should not be used by other packages.

Usage

lib(...)

Arguments

...
[any]
Package names.

Value

[logical]: Named logical vector determining the success of package load.

Examples

## Not run: 
##D lib("BBmisc", "MASS", "rpart")
## End(Not run)

[Package BBmisc version 1.6 Index]