lsort {BBmisc}	R Documentation

A wrapper for sort to sort using the “C” collating rules.

Description

A wrapper for sort to sort using the “C” collating rules.

Usage

lsort(...)

Arguments

...
Options passed to sort.

Value

See sort.

[Package BBmisc version 1.6 Index]