makeFileCache {BBmisc}	R Documentation

A caching wrapper around load2.

Description

This closure returns a wrapper around load2 which per default caches loaded objects and returns the cached version in subsequent calls.

Usage

makeFileCache(use.cache = TRUE)

Arguments

use.cache
[logical(1)]
Enable the cache? Default is TRUE.

Value

[function()] with argument slot (name of the slot to cache the object in, default is “default”). All other arguments are passed down to load2.

[Package BBmisc version 1.7 Index]