namedList {BBmisc}	R Documentation

Create named list, possibly initialized with a certain element.

Description

Create named list, possibly initialized with a certain element.

Usage

  namedList(names, init)

Arguments

names
[character]
Names of elements.

init
[valid R expression]
If given all list elements are initialized to this, otherwise NULL is used.

Value

[list].

Examples

namedList(c("a", "b"))

## $a
## NULL
## 
## $b
## NULL

namedList(c("a", "b"), init=1)

## $a
## [1] 1
## 
## $b
## [1] 1

[Package BBmisc version 1.5 Index]