save2 {BBmisc}	R Documentation

Save multiple objects to a file.

Description

A simple wrapper for save. Understands key = value syntax to save objects using arbitrary variable names. All options of save, except list and envir, are available and passed to save.

Usage

  save2(file, ..., ascii = FALSE, version = NULL,
    compress = !ascii, compression_level,
    eval.promises = TRUE, precheck = TRUE)

Arguments

file
File to save.

...
[any]
Will be converted to an environment and then passed to save.

ascii
See help of save.

version
See help of save.

compress
See help of save.

compression_level
See help of save.

eval.promises
See help of save.

precheck
See help of save.

Value

See help of save.

Examples

x <- 1
save2(y=x, file=tempfile())

[Package BBmisc version 1.5 Index]