setValue {BBmisc}	R Documentation

Set a list element to a new value.

Description

This wrapper supports setting elements to NULL.

Usage

setValue(obj, index, newval)

Arguments

obj
[list]

index
[character | integer]
Index or indices where to insert the new values.

newval
[any]
Inserted elements(s). Has to be a list if index is a vector.

Value

[list]

[Package BBmisc version 1.7 Index]