stopf {BBmisc}	R Documentation

Wrapper for stop and sprintf.

Description

A wrapper for stop with sprintf applied to the arguments. Notable difference is that error messages are not truncated to 1000 characters by default.

Usage

  stopf(..., warning.length = 8170L)

Arguments

...
[any]
See sprintf.

warning.length
[integer(1)]
Number of chars after which the error message gets truncated, see ?options. Default is 8170.

Value

Nothing.

Examples

err <- "an error."
try(stopf("This is %s", err))

[Package BBmisc version 1.5 Index]