toRangeStr {BBmisc}	R Documentation

Convert a numerical vector into a range string.

Description

Convert a numerical vector into a range string.

Usage

toRangeStr(x, range.sep = " - ", block.sep = ", ")

Arguments

x
[integer]
Vector to convert into a range string.

range.sep
[character(1)]
Separator between the first and last element of a range of consecutive elements in x. Default is “ - ”.

block.sep
[character(1)]
Separator between non consecutive elements of x or ranges. Default is “, ”.

Value

[character(1)]

Examples

x <- sample(1:10, 7)
toRangeStr(x)

## [1] "1 - 2, 4 - 5, 7 - 8, 10"

[Package BBmisc version 1.5 Index]