warningf {BBmisc}	R Documentation

Wrapper for warning and sprintf.

Description

A wrapper for warning with sprintf applied to the arguments.

Usage

warningf(..., immediate = TRUE, warning.length = 8170L)

Arguments

...
[any]
See sprintf.

immediate
[logical(1)]
See warning. Default is TRUE.

warning.length
[integer(1)]
Number of chars after which the warning message gets truncated, see ?options. Default is 8170.

Value

Nothing.

Examples

msg <- "a warning"
warningf("this is %s", msg)

## Warning: this is a warning

[Package BBmisc version 1.5 Index]