which.first {BBmisc}	R Documentation

Find the index of first/last TRUE value in a logical vector.

Description

Find the index of first/last TRUE value in a logical vector.

Usage

which.first(x, use.names = TRUE)

which.last(x, use.names = TRUE)

Arguments

x
[logical]
Logical vector.

use.names
[logical(1)]
If TRUE and x is named, the result is also named.

Value

[integer(1) | integer(0)]. Returns the index of the first/last TRUE value in x or an empty integer vector if none is found.

[Package BBmisc version 1.7 Index]