isScalarValue {BBmisc}	R Documentation

Is given argument an atomic vector or factor of length 1?

Description

Is given argument an atomic vector or factor of length 1?

Usage

isScalarValue(x, na.ok = TRUE, null.ok = FALSE)

Arguments

x
[any]
Argument.

na.ok
[logical(1)]
Is NA considered a scalar? Default is TRUE.

null.ok
[logical(1)]
Is NULL considered a scalar? Default is FALSE.

Value

[logical(1)].

[Package BBmisc version 1.6 Index]